Split weather component code
